requirements.txt
streamlit
pandas
pyyaml
requests
scikit-learn
openai 
